Move components to app/components
