+++ src/app/components/config/config.interface.ts
