+++ src/app/components/config/config.ts
