+++ src/app/components/figma/figma.interface.ts
