+++ src/app/components/platform/platform.interface.ts
