+++ src/app/components/platform/platform.service.ts
