+++ src/app/components/template/template.interface.ts
