+++ src/app/components/translator/translator.interface.ts
