+++ src/app/components/translator/translator.service.ts
